pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "polymath"
version = "0.2.1"
description = "An extensible AI Polymath powered by MCP architecture."
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "google-adk>=0.1.0",
    "google-generativeai>=0.8.0",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "duckduckgo-search>=6.0.0",
    "pillow>=10.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "trafilatura>=1.6.0"
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[project.scripts]
polymath = "src.host.cli:entry_point"
pl = "src.host.cli:entry_point"

[tool.setuptools.packages.find]